Add files via upload
This web app loads an image and then creates two color wheels, one based on the modern Red-Green-Blue system and the other on the traditional Red-Yellow-Blue system. They are shown with 24 color bins each, as opposed to the usual 12 colors. The relative intensities of the hues in the image reflect the portion of the image with each hue. 

Some diagrams of common "color harmonies" are along the right-hand edge, for the purposes of comparison with the color wheels. The color harmonies have been developed over time for the traditional RYB color wheel.

In addition, the image can be viewed as pure hues, posterized by the 24 different color hues, either in the RGB or RYB systems. A five-stop grayscale zone image, inspired by Ansel Adam's zone system, is also available.